VFB-42 Camel Case Key
