+++ src/pdf/ShoppingList/getShoppingListData.ts
@@ -1,83 +1,22 @@
 import { Schema } from "@/database_utils";
 import supabase from "@/supabaseClient";
+import { fetchClients, fetchComment, fetchFamilies } from "@/common/fetch";
 import {
-    fetchClients,
-    fetchComment,
-    fetchFamilies,
-    fetchLists,
-    fetchParcels,
-} from "@/common/fetch";
-import {
-    ClientSummary,
-    HouseholdSummary,
     prepareClientSummary,
-    prepareHouseholdSummary,
     prepareRequirementSummary,
     processExtraInformation,
-    RequirementSummary,
-} from "@/common/format";
-
-export interface ParcelInfo {
-    voucherNumber: string;
-    packingDate: string;
-    collectionDate: string;
-    collectionSite: string;
-}
-
-interface ParcelInfoAndClientID {
-    parcelInfo: ParcelInfo;
-    clientID: string;
-}
+} from "@/common/formatClientsData";
+import { prepareHouseholdSummary } from "@/common/formatFamiliesData";
+import { prepareParcelInfo } from "@/pdf/ShoppingList/getParcelsData";
+import {
+    prepareItemsList,
+    ShoppingListPDFDataProps,
+} from "@/pdf/ShoppingList/shoppingListPDFDataProps";
 
 interface ClientDataAndFamilyData {
     clientData: Schema["clients"];
     familyData: Schema["families"][];
 }
-
-export interface ShoppingListPDFDataProps {
-    postcode: string;
-    parcelInfo: ParcelInfo;
-    clientSummary: ClientSummary;
-    householdSummary: HouseholdSummary;
-    requirementSummary: RequirementSummary;
-    itemsList: Item[];
-    endNotes: string;
-}
-
-export interface Item {
-    description: string;
-    quantity: string;
-    notes: string;
-}
-
-const formatDate = (dateString: string | null): string => {
-    if (!dateString) {
-        return "";
-    }
-    return new Date(dateString).toLocaleString("en-GB", {
-        year: "numeric",
-        month: "numeric",
-        day: "numeric",
-        hour: "numeric",
-        minute: "numeric",
-    });
-};
-
-const prepareParcelInfo = async (parcelID: string): Promise<ParcelInfoAndClientID> => {
-    const fetchedData = await fetchParcels(parcelID, supabase);
-    const parcelInfo: ParcelInfo = {
-        voucherNumber: fetchedData.voucher_number ?? "",
-        packingDate: formatDate(fetchedData.packing_datetime),
-        collectionDate: formatDate(fetchedData.collection_datetime),
-        collectionSite: fetchedData.collection_centre ?? "",
-    };
-
-    if (parcelInfo.collectionSite === "Delivery") {
-        parcelInfo.collectionSite = "N/A - Delivery";
-    }
-
-    return { parcelInfo: parcelInfo, clientID: fetchedData.client_id };
-};
 
 const getClientAndFamilyData = async (clientID: string): Promise<ClientDataAndFamilyData> => {
     const clientData = await fetchClients(clientID, supabase);
@@ -86,32 +25,7 @@ const getClientAndFamilyData = async (clientID: string): Promise<ClientDataAndFa
     return { clientData: clientData, familyData: familyData };
 };
 
-const getQuantityAndNotes = (
-    row: Schema["lists"],
-    size: number
-): Pick<Item, "quantity" | "notes"> => {
-    if (size >= 10) {
-        size = 10;
-    }
-    const size_quantity = `${size}_quantity` as keyof Schema["lists"];
-    const size_notes = `${size}_notes` as keyof Schema["lists"];
-    return {
-        quantity: row[size_quantity]?.toString() ?? "",
-        notes: row[size_notes]?.toString() ?? "",
-    };
-};
-
-const prepareItemsList = async (householdSize: number): Promise<Item[]> => {
-    const listData = await fetchLists(supabase);
-    return listData.map((row): Item => {
-        return {
-            description: row.item_name,
-            ...getQuantityAndNotes(row, householdSize),
-        };
-    });
-};
-
-const prepareData = async (parcelID: string): Promise<ShoppingListPDFDataProps> => {
+const getShoppingListData = async (parcelID: string): Promise<ShoppingListPDFDataProps> => {
     const { parcelInfo, clientID } = await prepareParcelInfo(parcelID);
     const { clientData, familyData } = await getClientAndFamilyData(clientID);
     const itemsList = await prepareItemsList(familyData.length);
@@ -136,4 +50,4 @@ const prepareData = async (parcelID: string): Promise<ShoppingListPDFDataProps> 
     };
 };
 
-export default prepareData;
+export default getShoppingListData;
